+++ src/App.js
@@ -1,0 +1,6 @@
+import React from "react";
+import { createRoot } from "react-dom/client";
+import Main from "./Main.jsx";
+
+const root = createRoot(document.getElementById("root"));
+root.render(React.createElement(Main));
