+++ src/components/listbox/ListboxDelete.jsx
@@ -1,10 +1,10 @@
-import "./listbox.css";
+import "./Listbox.css";
 
-function ListboxDelete({ todo, setTodos }) {
+const ListboxDelete = ({ todo, setTodos }) => {
   const handleDelete = () => {
     setTodos((prevTodos) => prevTodos.filter((item) => item !== todo));
   };
   return <div className="delete" onClick={handleDelete} />;
-}
+};
 
 export default ListboxDelete;
