+++ src/components/textbox/TextboxAdd.jsx
@@ -1,7 +1,7 @@
-import "./textbox.css";
+import "./Textbox.css";
 
-function TextboxAdd({ setTodos, text, setText, handleSubmit }) {
+const TextboxAdd = ({ setTodos, text, setText, handleSubmit }) => {
   return <div className="input_button" onClick={handleSubmit} />;
-}
+};
 
 export default TextboxAdd;
